+++ wallet-test/src/test/scala/org/bitcoins/wallet/models/TransactionDAOTest.scala
@@ -1,9 +1,6 @@
 package org.bitcoins.wallet.models
 
-import org.bitcoins.core.wallet.transactions.{
-  TransactionDb,
-  TransactionDbHelper
-}
+import org.bitcoins.core.api.wallet.db.{TransactionDb, TransactionDbHelper}
 import org.bitcoins.testkit.fixtures.WalletDAOFixture
 import org.bitcoins.testkit.wallet.{BitcoinSWalletTest, WalletTestUtil}
