+++ wallet/src/main/scala/org/bitcoins/wallet/internal/TransactionProcessing.scala
@@ -1,6 +1,11 @@
 package org.bitcoins.wallet.internal
 
-import org.bitcoins.core.api.wallet.db.{AddressTagDb, SpendingInfoDb}
+import org.bitcoins.core.api.wallet.db.{
+  AddressTagDb,
+  SpendingInfoDb,
+  TransactionDb,
+  TransactionDbHelper
+}
 import org.bitcoins.core.currency.CurrencyUnit
 import org.bitcoins.core.number.UInt32
 import org.bitcoins.core.protocol.BitcoinAddress
@@ -8,11 +13,7 @@ import org.bitcoins.core.protocol.blockchain.Block
 import org.bitcoins.core.protocol.transaction.{Transaction, TransactionOutput}
 import org.bitcoins.core.util.FutureUtil
 import org.bitcoins.core.wallet.fee.FeeUnit
-import org.bitcoins.core.wallet.transactions.{
-  TransactionDb,
-  TransactionDbHelper
-}
-import org.bitcoins.core.wallet.utxo._
+import org.bitcoins.core.wallet.utxo.{AddressTag, TxoState}
 import org.bitcoins.crypto.{DoubleSha256Digest, DoubleSha256DigestBE}
 import org.bitcoins.wallet._
 import org.bitcoins.wallet.api.{AddUtxoError, AddUtxoSuccess}
